% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alignment_from_cigar}
\alias{alignment_from_cigar}
\title{Translate CIGAR string}
\usage{
alignment_from_cigar(bam_df, qname_input, genome)
}
\arguments{
\item{bam_df}{BAM DataFrame object}

\item{qname_input}{Read ID}

\item{genome}{Genome assembly in quotation marks}
}
\description{
This function reads the CIGAR string and other information of a specified alignment from the loaded
BAM file to visualize its alignment to the reference. It uses BSgenome.Hsapiens
libraries to load reference sequences.
}
\examples{
\dontrun{
data("loaded_Bam", package = "CIGARTranslate") # loads myBam
alignment_from_cigar(myBam, "A00491:61:HFCF7DMXX:1:1222:1389:1705", "hg38")
}
}
\references{
\url{https://en.wikipedia.org/wiki/Binary_Alignment_Map}\cr
}
\author{
Isidora Bijelovic\cr Politecnico di Milano\cr Maintainer: Isidora
Bijelovic\cr E-Mail: \href{mailto:isidora.bijelovic@mail.polimi.it}{isidora.bijelovic@mail.polimi.it}
}
