% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIGARTranslate-package.R
\docType{package}
\name{CIGARTranslate-package}
\alias{CIGARTranslate-package}
\alias{CIGARTranslate}
\title{CIGARTranslate}
\description{
The package \code{CIGARTranslate} provides functions used to visualize
the alignment of a mapped read from a BAM file.
}
\references{
\url{https://en.wikipedia.org/wiki/Binary_Alignment_Map}\cr
}
\author{
Isidora Bijelovic\cr Politecnico di Milano\cr Maintainer: Isidora
Bijelovic\cr E-Mail: \href{mailto:isidora.bijelovic@mail.polimi.it}{isidora.bijelovic@mail.polimi.it}
}
