% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaded_Bam.R
\docType{data}
\name{myBam}
\alias{myBam}
\alias{loaded_Bam}
\title{Example BAM for CIGARTranslate package}
\format{
A DFrame object representing a BAM
}
\usage{
data(loaded_Bam)
}
\description{
This dataset contains a portion of a BAM file to be used as an example for the CIGARTranslate package.
}
\seealso{
\code{\link{alignment_from_cigar}}
}
\keyword{datasets}
