% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_bam}
\alias{read_bam}
\title{Load BAM file}
\usage{
read_bam(bam_file_path)
}
\arguments{
\item{bam_file_path}{Path with forward slashes enclosed in quotation marks}
}
\value{
BAM DataFrame
}
\description{
This function uses Rsamtools::scanBam to read a user-specified BAM file and
converts it to a DataFrame object that is compatible with the next function.
}
\references{
\url{https://en.wikipedia.org/wiki/Binary_Alignment_Map}\cr
}
\author{
Isidora Bijelovic\cr Politecnico di Milano\cr Maintainer: Isidora
Bijelovic\cr E-Mail: \href{mailto:isidora.bijelovic@mail.polimi.it}{isidora.bijelovic@mail.polimi.it}
}
